==> notebooks/00_DWD_Data.ipynb <==
import os
import pandas as pd
import numpy as np

path=os.path.join(os.getcwd(), '../data/input/dresden/')
path

df = pd.read_csv(path+'df_dresden.csv')
df.describe()

df.info()

df.head()

df['datetime'] = pd.to_datetime(df['datetime'].astype(str), errors='coerce')
df.info()

df.sort_values(by=['datetime'])
df = df.drop(['Unnamed: 0'], axis = 1)

## Wetterdaten Station 01051 Dresden-Strehlen (relativ zentral)

- ACHTUNG: Missing values sind mit -999 markiert!
- air temperature
- precipitation

path=os.path.join(os.getcwd(), '../data/input/dwd/01051/')
path

df_air_temp_01051 = pd.read_csv(path+'air_temperature_20181108_20200510_01051.txt',sep=';')
df_prec_01051 = pd.read_csv(path+'precipitation_20181108_20200510_01051.txt',sep=';')

df_01051  = pd.merge(df_air_temp_01051, df_prec_01051, on="MESS_DATUM")
df_01051

# drop irrelevant columns
df_01051 = df_01051.drop(['STATIONS_ID_x', 'QN_9', 'eor_x', 'STATIONS_ID_y', 'QN_8', 'WRTR', 'eor_y'], axis = 1)
df_01051

# replace -999 with nan
df_01051 = df_01051.replace(-999,np.NaN)

df_01051.isna().sum()

## Wetterdaten Station 01050 Dresden-Hosterwitz (Südöstlich)

- air temperature
- precipitation
- sun

path=os.path.join(os.getcwd(), '../data/input/dwd/01050/')
path

df_air_temp_01050 = pd.read_csv(path+'air_temperature_20181108_20200510_01050.txt',sep=';')
df_prec_01050 = pd.read_csv(path+'precipitation_20181108_20200510_01050.txt',sep=';')
df_sun_01050 = pd.read_csv(path+'sun_20181108_20200510_01050.txt',sep=';')

df_01050 = pd.merge(df_air_temp_01050, df_prec_01050, on="MESS_DATUM")
df_01050 = pd.merge(df_01050, df_sun_01050, on="MESS_DATUM")
df_01050

# drop irrelevant columns
df_01050 = df_01050.drop(['STATIONS_ID_x', 'QN_9', 'eor_x', 'STATIONS_ID_y', 'QN_8', 'WRTR', 'eor_y', "STATIONS_ID",'QN_7', 'eor' ], axis = 1)
df_01050

df_01050 = df_01050.replace(-999,np.NaN)
df_01050.isna().sum()

## Wetterdaten Station 01048 Dresden-Klotsche (Flughafen)

- air temperature
- precipitation
- sun
- wind

path=os.path.join(os.getcwd(), '../data/input/dwd/01048/')
path

df_air_temp_01048 = pd.read_csv(path+'air_temperature_20181108_20200510_01048.txt',sep=';')
df_prec_01048 = pd.read_csv(path+'precipitation_20181108_20200510_01048.txt',sep=';')
df_sun_01048 = pd.read_csv(path+'sun_20181108_20200510_01048.txt',sep=';')
df_wind_01048 = pd.read_csv(path+'wind_20181108_20200510_01048.txt',sep=';')

df_01048 = pd.merge(df_air_temp_01048, df_prec_01048, on="MESS_DATUM")
df_01048 = pd.merge(df_01048, df_sun_01048, on="MESS_DATUM")
#drop irrelevant columns
df_01048 = df_01048.drop(['STATIONS_ID_x', 'QN_9', 'eor_x', 'STATIONS_ID_y', 'QN_8', 'WRTR', 'eor_y', "STATIONS_ID",'QN_7', 'eor'], axis = 1)
#continue merge
df_01048 = pd.merge(df_01048, df_wind_01048, on="MESS_DATUM")
df_01048.info()

# drop irrelevant columns
df_01048 = df_01048.drop(['STATIONS_ID', 'QN_3', '   D', 'eor'], axis = 1)
df_01048

df_01048 = df_01048.replace((-999),np.NaN)
df_01048.isna().sum()

## Finaler Wetterdatensatz

- Durchschnittswerte von air_temperature, precipitation, wind & sun
- falls kein Wert vorhanden, wird Wert bei Berechnung ignoriert
- irrelevante Spalten wurden im Vorraus verworfen

#merge mean of TT_TU, RF_TU, R1, RS_IND from all stations
df_dwd = pd.concat([df_01050.iloc[:,:5], df_01051.iloc[:,:5], df_01048.iloc[:,:5]]).groupby('MESS_DATUM', as_index=False).mean()
#merge mean of SD_SO from station 01048 & 01050
df_dwd_tmp = pd.concat([df_01048.iloc[:,[0,5]],df_01050.iloc[:,[0,5]]]).groupby('MESS_DATUM', as_index=False).mean()
#merge mean SD_SO
df_dwd = pd.merge(df_dwd, df_dwd_tmp, on="MESS_DATUM") 
#merge wind from 01048
df_dwd = pd.merge(df_dwd, df_01048.iloc[:,[0,6]], on="MESS_DATUM")
df_dwd

display(df_dwd.isna().sum(), df_dwd.describe(), df_dwd.info())
#vollständiger Datensatz ohne Nullwerte, da diese durch andere Wetterstationen 'gefüllt' worden sind

## Merge Wetterdatensatz with NextBike data

#Wetterdaten MESS_DATUM in datetime kovertieren
#datetime besitzt zustätzlich Angaben zu Minutne und Sekunden, daher fehlen vier Nullen
df_dwd['MESS_DATUM'] = df_dwd['MESS_DATUM'] * 10000
df_dwd['MESS_DATUM'] = pd.to_datetime(df_dwd['MESS_DATUM'].astype(str), errors='coerce')
df_dwd.info()
#... datetime done

display(df_dwd.tail(),df.head())

#(df_dwd["MESS_DATUM"]=="2019-01-20 01:00:00").sum()
(df_dwd["MESS_DATUM"]=="2019-01-20 02:00:00").sum()
#aus irgendeineinem grund fehlen einträge zu bestimmten Uhrzeiten? evtl fehler bei concat der wetterdaten? 
#mögliche lösung ist es, index zu resetten im vorraus

df_full = pd.merge_asof(df.sort_values('datetime'), df_dwd, left_on='datetime', right_on='MESS_DATUM', tolerance=pd.Timedelta('11 hour'),
                 allow_exact_matches=True, direction='nearest')


df_full.isna().sum()
# with timedelta of 11 hours we have zero NaNs for our weather data
# this suggests the method of concatination got rid of some incidents (maybe sun??)
# this means our weather data is not very reliable atm, the best time delta possible would be 30 mins
# especially for time frames around 12 pm


df_full
