==> notebooks/04_Merge_Hexagon_Chunks.ipynb <==
# Merge Hexagon Chunks

This notebook merges the created hexagon files to create a Trips_Hexagons.csv including all trips each with its corresponding hexagons.

### What is done?
1. Load hexagon chunk files created on Mac & merge them into single file
2. Save file on disk
    - file is already set on gitignore

### Input
- data/output/trips-hexagons/trips-hexagons-{id}.csv
    - (Multiple chunkfiles, each with antoher number as id)

### Output
- data/output/Trips_Hexagons.csv
    
---------------------------------------------------------

## 1. Load hexagon chunk files created on Mac & merge them into single file

import pandas as pd
import os
import glob

path_trips_hex_chunks=os.path.join(os.getcwd(), "..", "data", "output", "trips-hexagons")

print(path_trips_hex_chunks)

directors = glob.glob(path_trips_hex_chunks)

all_files = []
for director in directors:
    all_files=all_files+glob.glob(director+"/*.csv")
    
li = []

for filename in all_files:
    df_read = pd.read_csv(filename, index_col=None, header=0)
    li.append(df_read)

df = pd.concat(li, axis=0, ignore_index=True)

## 2. Save file on disk

if "Unnamed: 0" in df.columns:
    df = df.drop("Unnamed: 0", axis=1)

df

df.to_csv(os.path.join(os.getcwd(), "..", "data", "output", "Trips.csv"))
