==> notebooks/06_Availability.ipynb <==
# Availability Information

Adds information about idle bikes and available bikes to each trip in "Trips.csv".

## Disclaimer
As the calculations in this notebook throw out negative values for idle bikes and available bikes we assume that we have some mistake in our calculation.
To be able to show the systematics and understanding of the following training and prediction algorithms,
we decided to set random values in a range from [0 to 200] as idle bikes
and random values in a range from [0 to 500] as availability of bikes
With these random values, we will work from here on with the further algorithms.

## Input:
- "Trips.csv"
- "hexagons-big-dresden.csv"

## Output:
- Modified "Trips.csv"

import pandas as pd
import os
from random import randint

path_output=os.path.join(os.getcwd(), '..', 'data', 'output')
path_trips=os.path.join(path_output, "Trips.csv")
path_hex_big=os.path.join(path_output, "hexagons-big-dresden.csv")

print(path_trips)

df_trips = pd.read_csv(path_trips, index_col="Unnamed: 0")
df_hex_big = pd.read_csv(path_hex_big, index_col="Unnamed: 0")

display(df_trips,df_hex_big)

df_trips["datetime_start"] = pd.to_datetime(df_trips["datetime_start"])
df_trips["datetime_end"] = pd.to_datetime(df_trips["datetime_end"])
df_trips["24_agg_time"] = pd.to_datetime(df_trips["24_agg_time"])
df_trips["24_agg_time_end"] = pd.to_datetime(df_trips["24_agg_time_end"])
df_trips["1_agg_time"] = pd.to_datetime(df_trips["1_agg_time"])
df_trips.columns

# CONSTANTS
CUR_HEX = "861f1b097ffffff"
CUR_AGG_TIME = pd.to_datetime("2019-01-20")

def calculate_idle_available(df_trips, cur_hex, cur_agg_time):
    """
    # print("Calculating idle and available bikes for hexagon", cur_hex, "during aggregation time period", cur_agg_time)
    
    # === 1: CALCULATE IDLE BIKES ===
    # get all trips ending in current hexagon
    df_current_hex = df_trips[df_trips["h3_hex_big_id_end"] == cur_hex]
    # get all trips ending before beginning of aggregation time
    df_end_before_agg_time = df_current_hex[df_current_hex["datetime_end"] < cur_agg_time]
    # groupby bikenumber and get latest trip ending befor aggr time (= number of bikes that are in hexagon when aggr time begins)
    bikes_before = df_end_before_agg_time.groupby("b_number_start")["datetime_end"].max().count()
    
    # get all trips starting in current hexagon
    df_current_hex_start = df_trips[df_trips["h3_hex_big_id_start"] == cur_hex]
    # get all trips starting after aggregation time
    df_start_before_agg_time_end = df_current_hex_start[df_current_hex_start["24_agg_time"] == cur_agg_time]
    # count number of bikes started after aggregation time
    bikes_used_in_agg_time = df_start_before_agg_time_end.groupby("b_number_start")["datetime_start"].count().count()
    # print("bikes_used_in_agg_time",bikes_used_in_agg_time)
    
    # idle bikes = number of bikes when aggr time starts - number of bikes used after aggr time ends
    # (so inside aggr time period they were not used)
    idle_bikes = bikes_before - bikes_used_in_agg_time
    # print("Idle bikes:", idle_bikes)
    
    # === 2: CALCULATE AVAILABLE BIKES ===
    # get number of bikes started in current hexagon during current agg time
    df_trips_started = df_trips[(df_trips["24_agg_time"] == cur_agg_time) & (df_trips["h3_hex_big_id_start"] == cur_hex)]["24_demand_hex_big"]
    if (len(df_trips_started) == 0):
        trips_started = 0
    else:
        trips_started = df_trips_started.iloc[0]
    # print("Trips started (in current hexagon, during current agg time):",trips_started)
    # get number of trips ended in current hexagon during current agg time
    df_trips_ended = df_trips[(df_trips["24_agg_time_end"] == cur_agg_time) & (df_trips["h3_hex_big_id_end"] == cur_hex)]["24_demand_hex_big_end"]
    if (len(df_trips_ended) == 0):
        trips_ended = 0
    else:
        trips_ended = df_trips_ended.iloc[0]
    # print("Trips ended (in current hexagon, during current agg time):",trips_ended)
    
    # calculate available bikes for current hexagon & time period
    available_bikes = trips_ended - trips_started + idle_bikes
    # print("Available bikes:", available_bikes)
    """
    
    # """
    # === DISCLAIMER ===
    # As the calculations above throw out negative values for idle bikes and available bikes we assume that we have some mistake in our calculation.
    # To be able to show the systematics and understanding of the following training and prediction algorithms,
    # we decided to set random values in a range from [0 to 200] as idle bikes
    # and random values in a range from [0 to 500] as availability of bikes
    # With these random values, we will work from here on with the further algorithms.
    
    idle_bikes = randint(0,200)
    available_bikes = randint(0,500)
    # """
    
    # add newly calculated values to corresponding entries in trips data
    df_trips.loc[(df_trips["24_agg_time_end"] == cur_agg_time) & (df_trips["h3_hex_big_id_end"] == cur_hex),"24_idle_hex_big"]=idle_bikes
    df_trips.loc[(df_trips["24_agg_time_end"] == cur_agg_time) & (df_trips["h3_hex_big_id_end"] == cur_hex),"24_available_hex_big"]=available_bikes
    
    return df_trips

hex_big_ids = df_hex_big["h3_hex_big_id"].unique().tolist()
all_agg_times = df_trips["24_agg_time"].unique()
all_agg_times = pd.to_datetime(all_agg_times)
display(hex_big_ids, all_agg_times)

df_trips = calculate_idle_available(df_trips=df_trips, cur_hex=CUR_HEX, cur_agg_time=CUR_AGG_TIME)
df_trips[df_trips["24_agg_time_end"] == CUR_AGG_TIME]

for hex_id in hex_big_ids:
    print("Hexagon is:", hex_id)
    for agg_time in all_agg_times:
        df_trips = calculate_idle_available(df_trips=df_trips, cur_hex=hex_id, cur_agg_time=agg_time)

df_trips

print(df_trips["24_idle_hex_big"].isna().sum().sum())
print(df_trips["24_available_hex_big"].isna().sum().sum())

df_trips[df_trips["24_available_hex_big"] < 0]["24_agg_time_end"].unique()

avg_idle = int(df_trips["24_idle_hex_big"].mean())
df_trips["24_idle_hex_big"].fillna(avg_idle, inplace=True)
avg_available = int(df_trips["24_available_hex_big"].mean())
df_trips["24_available_hex_big"].fillna(avg_available, inplace=True)
print(df_trips["24_idle_hex_big"].isna().sum().sum())
print(df_trips["24_available_hex_big"].isna().sum().sum())

df_trips.info()

df_trips.describe()

df_trips.to_csv(os.path.join(path_output, "Trips.csv"))
