==> notebooks/10_Create_Parameter_File.ipynb <==
import pandas as pd
import os

dict_params = {
    "test_size":0.3,
    "random_state":int(42),
    "epochs":20,
    "validation_size":0.2/0.7,
    "pca_components":20
}
sr_params = pd.Series(dict_params)
sr_params

sr_params.to_csv(os.path.join(os.getcwd(), "..", "data", "input", "params.csv"), index_label=["param", "value"])

pd.read_csv(os.path.join(os.getcwd(), "..", "data", "input", "params.csv")).drop("0", axis=1)
