==> notebooks/11_Feature_Creation.ipynb <==
# Feature preparation
Right here the features for the ml models are created and filtered. Then a csv with only the important features is saved.

import os
import pandas as pd
import pickle

from sklearn.preprocessing import LabelEncoder  
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import StandardScaler
from sklearn.decomposition import PCA

path_trips = os.path.join(os.getcwd(), "..", "data", "output", "Trips.csv")

### Take Params from csv

params = pd.read_csv(os.path.join(os.getcwd(), "..", "data", "input", "params.csv")).drop("0", axis=1)
_test_size = params[params["param"]=="test_size"]["value"].values[0]
_random_state = int(params[params["param"]=="random_state"]["value"].values[0])
_pca_components = int(params[params["param"]=="pca_components"]["value"].values[0])

df = pd.read_csv(path_trips)
df.head()

df.info()

df.isnull().sum()

## Drop not needed columns
Drop end information and not usefull columns.
The features are cleaned in three steps. <br>
- features_1 => drop not usefull columns 
- features_2 => drop end information
- features_3 => drop object types

print("Drop not int columns and not usefull")
# Unnamed: 0 => index column
# p_terminal_type_end => NaN values
features_1 = df.drop(["Unnamed: 0", 
                      #"p_terminal_type_start", 
                      "p_number_start", 
                      #"b_pedelec_battery_start"
                     ], axis=1)
print("Drop end information")
features_2 = features_1.drop(
    ["p_bike_racks_end",
     "p_spot_end", 
     "p_booked_bikes_end", 
     "p_place_type_end", 
     "datetime_end",
     "p_uid_end",
     "p_bikes_end",     
     "p_lat_end",
     "p_name_end",
     "p_free_racks_end",
     #"p_address_end",
     "p_number_end",
     "p_lng_end",
     "p_maintenance_end",
     "h3_hex_small_id_end",
     "h3_hex_big_id_end",
     "trip_duration",
     "24_demand_end",
     "24_demand_hex_small_end",
     "24_demand_hex_big_end",
     "24_agg_time_end",
     "6_demand_end",
     "6_demand_hex_big_end",
     "6_demand_hex_small_end",
     "6_agg_time_end"
     #"p_terminal_type_end", 
     #"p_bike_types_end"
    ], 
    axis=1
)
print("Drop object types...")
features_3 = features_2.drop(["datetime_start", 
                              "p_name_start", 
                              #"p_address_start", 
                              "b_lock_types_start",
                              "booking_date_start",
                              #"p_bike_types_start", 
                              #"MESS_DATUM"
                             ], axis=1)

### Dummy
Create Dummy variables for all booleans

p_spot_start = features_3["p_spot_start"].astype(int)
b_electric_lock_start = features_3["b_electric_lock_start"].astype(int)
p_maintenance_start = features_3["p_maintenance_start"].astype(int)
weekend = features_3["weekend"].astype(int)

le = LabelEncoder()
hexa_small = pd.Series(le.fit_transform(features_3["h3_hex_small_id_start"]))
hexa_big = pd.Series(le.fit_transform(features_3["h3_hex_big_id_start"]))

# weekend = p_df["Weekend"].astype(int)
features = features_3.drop(["p_spot_start", "b_electric_lock_start", "p_maintenance_start", "weekend", "h3_hex_small_id_start", "h3_hex_big_id_start"], axis=1)
features = pd.concat([features, p_spot_start, b_electric_lock_start, p_maintenance_start, weekend, hexa_small, hexa_big], axis=1)
features = features.rename({0:"hexa_small", 1:"hexa_big"}, axis=1)

### Feature Engineering

# features["XYZ"] = np.square(features["XXX"])

# Save features
features.to_csv(os.path.join(os.getcwd(), "..", "data", "output", "Features.csv"))

### Scaling and PCA

def train_scaler_pca(hex_size="hexa_small"):
    """
    Train Scaler and PCA depending on hex size.
    
    Args:
        hex_size (str): which hex size should be used? small or big possible
        
    Returns:
        No return
    """
    
    #FILTER
    features_X = features.drop(["24_demand", "24_demand_hex_big", "24_demand_hex_small", "24_agg_time",
                                "6_demand", "6_demand_hex_big", "6_demand_hex_small", "6_agg_time",
                                "2_demand", "2_demand_hex_big", "2_demand_hex_small", "2_agg_time",
                                "1_demand", "1_demand_hex_big", "1_demand_hex_small", "1_agg_time",
                                "24_available_hex_big"], axis=1)
    if hex_size=="hexa_small":
        features_X = features_X.drop("hexa_big", axis=1)
    else:
        features_X = features_X.drop("hexa_small", axis=1)
    
    features_y = features["24_demand"]
    
    #SPLIT
    X_train, X_test, y_train, y_test = train_test_split(features_X, features_y, random_state=_random_state, test_size=_test_size)

    # STANDARD SCALER
    st_scaler = StandardScaler()
    # fit scaler only on training set not on test set
    st_scaler.fit(X_train)

    # Save Scaler Object
    obj = st_scaler
    filename = "Standard_Scaler_"+hex_size+".pkl"
    pickle.dump(obj, open(os.path.join(os.getcwd(), "..", "data", "output", "models", filename), "wb"))
    #Scale
    X_train_scaled = st_scaler.transform(X_train)

    # PCA
    pca = PCA(n_components=_pca_components)
    pca.fit(X_train_scaled)
    pca_explained_variance = pca.explained_variance_ratio_
    print("Var explained:", pca_explained_variance)
    print("Sum var explained", sum(pca_explained_variance))

    # Save PCA Object
    obj = pca
    filename = "PCA_"+hex_size+".pkl"
    pickle.dump(obj, open(os.path.join(os.getcwd(), "..", "data", "output", "models", filename), "wb"))
    
    print("Done")

hex_sizes = ["hexa_small", "hexa_big"]
for size in hex_sizes:
    train_scaler_pca(size)
