==> notebooks/13_Predict_SVM.ipynb <==
# SVM Pediction
Predict the different time resolutions with a trained SVM.

import pandas as pd
import matplotlib.pyplot as plt
import pickle
import os
from sklearn.model_selection import train_test_split
from sklearn.metrics import r2_score

path_output=os.path.join(os.getcwd(), "..", "data", "output")

features = pd.read_csv(os.path.join(path_output, "Features.csv"))
features = features.drop("Unnamed: 0", axis=1)

### Take Params from csv

params = pd.read_csv(os.path.join(os.getcwd(), "..", "data", "input", "params.csv")).drop("0", axis=1)
_test_size = params[params["param"]=="test_size"]["value"].values[0]
_random_state = int(params[params["param"]=="random_state"]["value"].values[0])

# Prediction

def predict(time_res="24_demand", hex_size="hexa_big", kernel="linear"):
    """
        Predict the number of trips in a given time resolution.
        
        Args:
            time_res (str): time resolution to train on
            
        Returns:
            y_test_predicted (Series): Predicted test data
    
    """
    print("Kernel is", kernel)
    # Temporal Resolution
    print("Temporal Resolution is", time_res)
    # features_X = features.drop(["24_sum", "6_sum", "2_sum", "1_sum"], axis=1)
    features_X = features.drop(["24_demand", "24_demand_hex_big", "24_demand_hex_small", "24_agg_time",
                                "6_demand", "6_demand_hex_big", "6_demand_hex_small", "6_agg_time",
                                "2_demand", "2_demand_hex_big", "2_demand_hex_small", "2_agg_time",
                                "1_demand", "1_demand_hex_big", "1_demand_hex_small", "1_agg_time",
                                "24_available_hex_big"], axis=1)
    features_y = features[time_res]
    
    # Spatial Resolution
    print("Spatial Resolution is", hex_size)
    if hex_size=="hexa_small":
        features_X = features_X.drop("hexa_big", axis=1)
    else:
        features_X = features_X.drop("hexa_small", axis=1)
        
    print("Split Data with random state", _random_state, "and test size", str(_test_size)+"...")
    X_train, X_test, y_train, y_test = train_test_split(features_X, features_y, random_state=_random_state, test_size=_test_size)
    
    print("Scale Data with Standard Scaler...")
    with open(os.path.join(path_output, "models", "Standard_Scaler_"+hex_size+".pkl"), "rb") as f:
        standard_scaler = pickle.load(f)
    X_test_scaled = standard_scaler.transform(X_test)

    print("Do PCA on Data...")
    with open(os.path.join(path_output, "models", "PCA_"+hex_size+".pkl"), "rb") as f:
        pca = pickle.load(f)
    X_test_transformed = pca.transform(X_test_scaled)
    
    print("Load Model", "("+time_res+", "+hex_size+", "+kernel+")...")
    with open(os.path.join(path_output, "models", "SVR_model_"+time_res+"_"+hex_size+"_"+kernel+".pkl"), "rb") as f:
        svr_model = pickle.load(f)
    print("Predict...")
    y_test_predicted = svr_model.predict(X_test_transformed)
    
    return y_test_predicted, y_test

# time_resolution is interchangable with ["24_demand", "6_demand", "2_demand", "1_demand"]
# spatial_resolution is interchangable with ["hexa_small", "hexa_big"]
# kernel is interchangable with ["linear", "rbf", "poly"]
time_resolution = "24_demand"
spatial_resolution = "hexa_big"
kernel = "linear"

y_test_predicted, y_test = predict(time_res=time_resolution, hex_size=spatial_resolution, kernel=kernel)

print("Plot difference between Real and Predicted...")
fig, ax = plt.subplots(figsize=(16, 8), dpi=300)
ax.plot(y_test_predicted, y_test, "bo")
ax.set_title("Real Y vs Predicted Y ("+time_resolution+", "+spatial_resolution+", "+kernel+")", fontsize=20)
ax.set_xlabel("Predicted Y", fontsize=18)
ax.set_ylabel("Real Y", fontsize=18)
fig.savefig(os.path.join(path_output, "SVM_Real_vs_Predicted_"+time_resolution+"_"+spatial_resolution+"_"+kernel+".png"))
plt.close(fig)
r2 = r2_score(y_test, y_test_predicted)
print("R2:",r2)
print("Done")
